backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.6
PyJWT>=2.8.0
werkzeug>=3.0.0
python-dotenv>=1.0.0
pymongo>=4.6.0
eventlet>=0.35.0
Pillow>=10.0.0